requirements.txt
tensorflow
numpy
pillow
flask
cs50
werkzeug
pytz
setuptools